reorganize & refactor banners
