--- src/components/LoginDialog/index.js
@@ -1,9 +1,0 @@
-import React from 'react'
-import Dialog from '@santiment-network/ui/Dialog'
-import AnonBanner from '../AnonBanner/AnonBanner'
-
-export default props => (
-  <Dialog {...props}>
-    <AnonBanner />
-  </Dialog>
-)
