+++ src/components/banners/feature/InlineBanner.module.scss
@@ -9,7 +9,7 @@
   border: 1px solid var(--porcelain);
   border-radius: 4px;
   box-shadow: 0 6px 50px rgba(24, 27, 43, 0.03);
-  background-color: $white;
+  background-color: var(--white);
   background-image: url(./wavesWhite.svg);
   background-repeat: no-repeat;
   background-position: 115% 35%;
