+++ src/ducks/Studio/PremiumBanner/PremiumBanner.module.scss
@@ -18,7 +18,7 @@
   margin-right: 12px;
   border-radius: 50%;
   min-width: 24px;
-  min-height: 24px;
+  height: 24px;
   background: var(--texas-rose-light);
   display: flex;
   align-items: center;
